+++ python/GafferSceneUI/ConstraintUI.py
@@ -82,6 +82,7 @@ Gaffer.Metadata.registerNode(
 			""",
 
 			"plugValueWidget:type", "GafferSceneUI.ScenePathPlugValueWidget",
+			"scenePathPlugValueWidget:scene", "targetScene in",
 
 		],